Add files via upload
The video Downloader and Streamer is now online. The only issue so far is the 
requestVideo() in Client 
For now I have hardcoded it to 
Thread requestVideo = new Thread(new VideoDownloader("10.0.0.1","papa.mp4")); 
but you will need to supply the ip address of the guy who has the video and the video name.